% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIndices.R
\name{addIndices}
\alias{addIndices}
\title{Add common soil and vegetative indices to Sentinel 2}
\usage{
addIndices(s2Img)
}
\arguments{
\item{s2Img}{Sentinel 2 ImageCollection to use.}
}
\value{
The Image Collection with the indices for each image in the series
}
\description{
Adds common soil and vegetative indices to sentinel 2A image collections. This should be conducted
after applying the s2Mask function to clean the images.
}
\author{
Trevan Flynn
}
