% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPICP.R
\name{calcPICP}
\alias{calcPICP}
\title{Prediction interval coverage probability}
\usage{
calcPICP(models, eval, label)
}
\arguments{
\item{models}{The results from regScale() or regBoot().}

\item{eval}{Validation data in as a Feature.}
}
\value{
Returns a list with a plot of PICP to confidence intervals, the results and
Lins concordance coefficient (how well the plot is in a 1:1 line).
}
\description{
A function to calculate the PICP of bootstrap ee models. It is meant to evaluate
uncertainties of the predictions. It is basically a complete replica of the code
by Malone et al., (2011), we just do most of the calculations on images first.
}
\references{
B.P. Malone, A.B. McBratney, B. Minasny, (2011). Empirical estimates of uncertainty for mapping continuous depth functions of soil attributes. Geoderma, Volume 160, Issues 3–4,Pages 614-626. https://doi.org/10.1016/j.geoderma.2010.11.013.
}
\author{
Trevan Flynn
}
