% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catBoot.R
\name{catBoot}
\alias{catBoot}
\title{Function to bootstrap GEE classifiers}
\usage{
catBoot(covar, response, label, model = NULL, boot = 10, ratio = 0.6, ...)
}
\arguments{
\item{covar}{The covariate Image.}

\item{response}{The FeatureCollection with what needs to be predicted.}

\item{label}{Character of the response to predict.}

\item{model}{Character of which ee model to use.}

\item{boot}{Integer of how many bootstraps.}

\item{ratio}{The ration of subsampling to use in each bootstrap.}
}
\value{
A list of the modal predictions, all preds, all models and the training and
evaluation samples for each bootstrap.
}
\description{
A function to bootstrap ee models which is useful for small datasets and gives
a more robust measure of uncertainty. It also gives a better estimate of variable
importance.
}
\author{
Trevan Flynn
}
