% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catEval.R
\name{catEval}
\alias{catEval}
\title{Evaluation stats for categorical data}
\usage{
catEval(models, eval, response)
}
\arguments{
\item{models}{The model results from catScale() or catBoot().}

\item{eval}{A FeatureCollection to evaluate the model.}

\item{response}{The response of the model (e.g., "texture").}
}
\value{
Returns a list containing the kappa coefficient, accuracy, consumer's
accuracy (CA) and producer's accuracy (PA).
}
\description{
Evaluate categorical results from a bootstrapped ee model through rgee functions
If you would like more options look into the yardstick package.
}
\author{
Trevan Flynn
}
