% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certCalc.R
\name{certCalc}
\alias{certCalc}
\title{Calculate uncertainties from bootstrap ee models}
\usage{
certCalc(models, uncert = "range", limits = c(5, 95))
}
\arguments{
\item{models}{Are the results from rafikReg() or rafikBootReg().}

\item{uncert}{Is the type of uncertainty one wants calculated. Either "range" (prediction range),
"stdev" (standard deviation), "var" (variance) or "se" (standard error).}

\item{limits}{A vector of confidence intervals (defaults to 5 and 95)}
}
\value{
An Image of the results
}
\description{
Calculates the uncertainties from a bootstrap regression from rafikReg() or rafikBootReg().
}
\author{
Trevan Flynn
}
