% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciCalc.R
\name{ciCalc}
\alias{ciCalc}
\title{Calculate a confusion index (CI)}
\usage{
ciCalc(probs)
}
\arguments{
\item{probs}{Is the output from the probCalc() function.}
}
\value{
An image of the confusion index.
}
\description{
Calculates the spatial CI from a bootstrapped ee model from Burrough et al. (1997)
where CI = 1 - (Pm - pm-1).Pm is the highest probability where Pm-1 is the second highest probability.
The final CI is the median of both Images. This function calculates the median values.
}
\references{
Burrough, P.A., Van Gaans, P.F.M., Hootsmans, R., (1997). Continuous classification in soil survey: spatial correlation, confusion and boundaries. Geoderma 77, 115–135.
https://doi.org/10.1016/S0016-7061(97)00018-9.
}
\author{
Trevan Flynn
}
