% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaRast.R
\name{eaRast}
\alias{eaRast}
\title{A method to run equal-area splines on a stack or rasters}
\usage{
eaRast(x, spc, depths, var, type = "integrate")
}
\arguments{
\item{x}{Is the stack of spatrasters to be used to reduce.}

\item{spc}{A soilProfileCollection created from the spatraster to reduce.}

\item{depths}{Is a vector of the depths (cm) to integrate (between 1:200).}

\item{var}{Character of what property the spline should be run on.}

\item{type}{A character specifying the type of calculations to perform on the spline.
can either be "integrate" (cumulative sum), "sum" (descrete sum), "mean",
or "median".}
}
\value{
A list consisting of the reduced spatraster and all the elements
produced from the ea_spline function from the ithir R package (Malone et al., 2009).
}
\description{
This algorithm takes a stack of rasters or spatrasters (decreasing depths),
and runs an ea_spline (Malone et al., 2019) at each pixel. To get the final
value it then integrates the spline (e.g., carbon stock), which is much more
accurate than summing 1 cm intervals.
}
\references{
https://10.1016/S0016-7061(99)00003-8

http://dx.doi.org/10.1016/j.geoderma.2009.10.007.
}
