% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalSpline.R
\name{evalSpline}
\alias{evalSpline}
\title{Classify and evaluate catSpline}
\usage{
evalSpline(spl)
}
\arguments{
\item{spl}{is the spline output.}
}
\value{
a list from the raw probabilities but classified
}
\description{
Takes the output of catSpline() and puts out categorical evaluation
statistics from the observed layers to the fitted spline.
}
\author{
Trevan Flynn
}
