% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{global_data}
\alias{global_data}
\title{WoSIS harmonised soil texture classes}
\format{
## `global_data`
List with two sf layers:
\describe{
  \item{Harmonised}{The harmonised soil texture classes}
  \item{Uncertainties}{The uncertainties associated with each profiles and depth}
}
}
\source{
<https://www.isric.org/explore/isric-soil-data-hub>
}
\usage{
global_data
}
\description{
Global data of harmonised USDA soil texture classes of 101,610 profiles with their
uncertainties in a Geopackage. Hormonised to GlobalSoilMap standardised depths.
}
\keyword{datasets}
