% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impCalc.R
\name{impCalc}
\alias{impCalc}
\title{Calculate importance of bootstrapped ee models}
\usage{
impCalc(models, mn)
}
\arguments{
\item{models}{Is the results from a bootstrapped ee model (regScale(), catScale(), regBoot() or catBoot()).}

\item{mn}{The number of covariates to display in the plot (all others will be in the returned value).}
}
\value{
A list of [[1]] a ggplot2 lollipop plot and [[2]] results of importance in a tibble.
}
\description{
Code to get the average variable importance from a bootstrapped model in GEE.
}
\author{
Trevan Flynn
}
