% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Neno}
\alias{Neno}
\title{WoSIS data}
\format{
## `Neno`
A data frame with 212 rows and 4 columns:
\describe{
  \item{id}{Profile number or order, should align with a shapefile}
  \item{top}{top depth of a horizon in the profile}
  \item{botoom}{bottom depth of a horizon in the profile}
  \item{Class}{USDA soil texture classes}
}
}
\source{
<https://www.isric.org/explore/isric-soil-data-hub>
}
\usage{
Neno
}
\description{
A small data frame of a subset WoSIS data in Southern Malawi. The format is how
it should be if problems are experienced when using the catSpline function.
}
\keyword{datasets}
