% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogcCalc.R
\name{ogcCalc}
\alias{ogcCalc}
\title{Oblique geographic coordinates (OGC) through GEE}
\usage{
ogcCalc(area, crs = "EPSG: 3857", scale = 250, nr = 6)
}
\arguments{
\item{area}{A Feature, FeatureCollection or Geometry of the area of interest.}

\item{crs}{The coordinate system to use (best in projected) but not required.}

\item{scale}{The resolution to use.}

\item{nr}{The number of angles to produce.}
}
\value{
An image with each band representing the angles calculated
}
\description{
An algorithm to calculate OGC in GEE originally developed by Møller et al. (2020).
This algorithm is largly taken from them, however, this algorithm does not use any
raster image as an input.It also allows to adjust the resolution and coordinate system accordingly.
We have found coarse coordinates as a covarate gives artifacts and smoothing helps this.
}
\references{
Møller, A. B., Beucher, A. M., and Pouladi, N., Greve, M. H. (2020) Oblique geographic  coordinates as covariates for digital soil mapping, SOIL 6(2) 269 - 289https:/10.5194/soil-6-269-2020
}
\author{
Trevan Flynn
}
