% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phySpline.R
\name{phySpline}
\alias{phySpline}
\title{analytically Solvable Physics-informed Categorical Depth Function}
\usage{
phySpline(
  df,
  class.var,
  lam = 0.1,
  alpha = 0.1,
  d = c(0, 5, 15, 30, 60, 100, 200),
  order = levels(df[[class.var]]),
  gthreshold = 30,
  vlow = NULL,
  vhigh = NULL,
  mode = c("classification", "continuous"),
  show.progress = TRUE
)
}
\arguments{
\item{df}{a data frame with at least 4 columns with the first 3 named c("id","top","bottom"). The id column should have at least}

\item{class.var}{the character of the class to harmonise}

\item{lam}{the smoothing term [0, 1]}

\item{alpha}{the physics-informed term [0, 1]}

\item{d}{the vector to harmonise the soil depths to}

\item{order}{the latent ordering to map the classes to (default = class level order)}

\item{gthreshold}{the threshold of which the physics-informed loss function will start being applied}

\item{mode}{either "classification" or "continuous" for classification class.var must be a factor for now}
}
\value{
A list with 3 harmonised data frames with the classes, latent classes and uncertainties for each interval, a data frame of observations vs interpolations
and the lookup table of the data. It also returns the coefficients to be use in phySpline_Interp() as the function does not work off of 1 cm increments.
}
\description{
Physics-informed spline that is fully analytically solvable and continuous in its 3 loss functions (data fidelity/mass-preservation, smoothing term, and physics term)
will interpolate "unseen" transition classes and not just the modal class. Informs class behavior and only assigns classes after all calculations have been performed.
Can run splines on a massive amount of soil data relatively fast and extract information that is scientific, not just algorithmic. !!BETA VERSION!!
}
\references{
Flynn, T., Rasaei, Z., Kostecki, R., (2025). Bayes’ Inference and Nearest Neighbor Splines to Harmonise Soil Texture Classes to International Depth Standards. https://doi.org/10.2139/ssrn.5078983
}
\author{
Trevan Flynn
}
