% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phySpline_Interp.R
\name{phySpline_Interp}
\alias{phySpline_Interp}
\title{Interpolate phySplines}
\usage{
phySpline_Interp(
  depths = seq(0, 160, by = 1),
  coeffs,
  lookup = NULL,
  vlow = NULL,
  vhigh = NULL
)
}
\arguments{
\item{depths}{a vector of the depths to interpolate on to.}

\item{coeffs}{the coefficients for the profile returned from phySpline().}

\item{vlow}{the lowest value to interpolate}

\item{vhigh}{the highest value to interpolate}

\item{looup}{for categorical splines, the lookup table used from phySplines().}
}
\value{
a dataframe with the depths, the value per depth and optionally the class.
}
\description{
A helper function to interpolate values from the phySpline() function for visualisation purposes from
the coefficients and optional classes. Can interpolate categorical or continuous phySplines.
}
\author{
Trevan Flynn
}
