% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probCalc.R
\name{probCalc}
\alias{probCalc}
\title{Function to calculate probabilities}
\usage{
probCalc(models, covar, levels)
}
\arguments{
\item{models}{The results from rafiki() or rafikBoot().}

\item{covar}{The covariates used in the boostrapped models.}

\item{levels}{The number of classes being predicted.}
}
\value{
A list of probabilities for each bootstrap..
}
\description{
Obtain the probabilities from rafiki() or rafikboot() algorithms. This function
also gets the data ready for a confusion index.
}
\author{
Trevan Flynn
}
