% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafikBoot.R
\name{rafikBoot}
\alias{rafikBoot}
\title{Bootstapped Google Earth Engine classifiers}
\usage{
rafikBoot(
  x,
  y,
  label,
  model,
  mode = "classification",
  fun = "mode",
  boot,
  fact,
  scale = y$projection()$nominalScale(),
  ...
)
}
\arguments{
\item{x}{Is the FeatureCollection with the sampling observations and labels.}

\item{y}{Is the Image of the covariates, where each band is a covariate.}

\item{label}{A character of what is to be predicted.}

\item{model}{Character of which model to use.}

\item{mode}{Should it be a classification, multiprobability, binomial probability or regression}

\item{fun}{Which aggregation function to use.}

\item{boot}{Integer of how many bootstrap iterations to perform.}

\item{fact}{The fraction of subsamples to perform.}

\item{scale}{What scale to perform the calculations in.}

\item{...}{Additional arguments for the algorithm parameters (e.g., numberOfTrees)}
}
\description{
Bootstrapped Google Earth Engine classification models with either predifined
aggregation functions, custom functions or raw predictions.
}
\author{
Trevan Flynn
}
