% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafikCI.R
\name{rafikCI}
\alias{rafikCI}
\title{Confusion index (CI) for ee Models}
\usage{
rafikCI(probs)
}
\arguments{
\item{probs}{Probabilities from running setOutputMode("MULTIPROBABILITY") to
an ee$Classifier.}
}
\value{
A CI image.
}
\description{
Calculate a confusion index for an ee$Classifier through Burrough et al. 1997.
Unlike the ciCalc function this calculates a CI for a single model.
}
\author{
Trevan Flynn
}
