% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafikisol.R
\name{rafikisol}
\alias{rafikisol}
\title{Downscale categorical maps through GEE}
\usage{
rafikisol(
  covar,
  response,
  label = response$bandNames()$getInfo(),
  scale = covar$projection()$nominalScale(),
  model = NULL,
  stype = "r",
  region = covar$geometry(),
  sampNum = 100,
  boot = 10,
  ...
)
}
\arguments{
\item{covar}{The covariates to be used in the model (should be fine resolution).}

\item{response}{The image with the data to be sampled on (should be coarse resolution).}

\item{label}{Character of response being downloaded, default = bandNames().}

\item{scale}{The scale at which to perform analysis. Defaults to scale of covar.}

\item{model}{A character of which ee model to be used being random forest ("rf"),
gradient tree boost ("gtb"), cart ("cart") or support vector machines ("svm").}

\item{stype}{Character of sampling type to perform either stratified ("s") or random ("r").}

\item{sampNum}{The number of samples for each bootstrap. For "s", its the amount per strata.}

\item{boot}{An integer of the number of bootstraps to perform.}

\item{...}{Parameters to add to the chosen ee model (e.g., numberOfTrees).}
}
\value{
Returns a list of four with the mean predictions, all predictions,
all models and all samples of the bootstraps.
}
\description{
A function to downscale categorical soil maps with ee models through Monte Carlo sampling.
The algorithm is basically the same as the dissever algorithm by
Malone et al. (2012) and Roudier et al. (2017) besides the fact it is for categorical data
and utilizes GEE and therefore, enables parallel processing and lowers computational power.
}
\references{
Brendan P. Malone, Alex B. McBratney, Budiman Minasny, Ichsani Wheeler, (2012). A general method for downscaling earth resource information. Computers & Geosciences, Volume 41, Pages 119-125. https://doi.org/10.1016/j.cageo.2011.08.021.

P. Roudier, B.P. Malone, C.B. Hedley, B. Minasny, A.B. McBratney, (2017). Comparison of regression methods for spatial downscaling of soil organic carbon stocks maps. Computers and Electronics in Agriculture. Volume 142, Part A, Pages 91-100. https://doi.org/10.1016/j.compag.2017.08.021.
}
\author{
Trevan Flynn
}
