% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2Mask.R
\name{s2Mask}
\alias{s2Mask}
\title{Cloud and shadow mask for Sentinel 2A surface reflectance images from Google Earth Engine.}
\usage{
s2Mask(s2Img)
}
\arguments{
\item{s2Img}{The Sentinel 2A SR ImageCollection to clean.}
}
\value{
An ImageCollection with clouds and shadows masked.
}
\description{
Masks clouds and shadows as well as some angle corrections for each image of a Sentinel 2A image collection.
The function has no parameters.
}
\author{
Trevan Flynn
}
