% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsCreate.R
\name{tsCreate}
\alias{tsCreate}
\title{Create a time series that can be plotted in R from an image collection}
\usage{
tsCreate(
  imgCol,
  reducer = "mean",
  scale,
  geom,
  dateFormat = "YYYY-MM-dd",
  sf = FALSE
)
}
\arguments{
\item{imgCol}{ImageCollection to reduce (e.g., Sentinel time series).}

\item{reducer}{The reducer to be used. Either a character or a built reducer (e.g., ee$Reducer$mean()).}

\item{scale}{Scale at which to reduce values. Defaults to image$projection()$nominalScale().}

\item{geom}{Point or polygon of location to be summarised.}

\item{dateFormat}{Date format you want the dataframe to come out as.}

\item{sf}{Logical, should the data frame be converted to an sf object into local space.}
}
\value{
A feature with values of image for each date in time series.
}
\description{
Reduces the image collection into a feature which can be transformed into an sf object
to plot in R and for further analysis. Please note that geometry must be labeled geometry and that
it reduces the images based on the median.
}
\author{
Trevan Flynn
}
