% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catSpline.R
\name{catSpline}
\alias{catSpline}
\title{Function for categorical splines}
\usage{
catSpline(
  obj,
  class.var,
  type = "class",
  beta = 1,
  lam = 0.1,
  d = c(0, 5, 15, 30, 60, 100, 200),
  vlow = 0,
  vhigh = 1,
  show.progress = TRUE
)
}
\arguments{
\item{obj}{is a data frame with columns of profile top of horizon and bottom of horizon
These columns must be named id, top and bottom respectively to be turned into a spc.f}

\item{class.var}{is the name of the factor to run the spline.}

\item{type}{type of output being "class" with the classification or "probs" with the raw probabilities.}

\item{beta}{is how smooth the softmedian value is where lower is smoother and
higher is closer to the true median.}

\item{lam}{is the smoothing factor of the spline}

\item{d}{is the depths to harmonise to.}

\item{vlow}{is the lower boundary value (lowest = 0).}

\item{vhigh}{is the higher boundary value (highest = 0).}
}
\value{
a list of harmonised posterior probabilites, observed vs predictions,
predictions in 1 cm increments (for figures) and the lookup table with probabilities.
}
\description{
Function runs equal-area splines on the probabilities of soil categorical data
and then classifies the soil class through a 1D nearest neighbor approach. Based of
the code written by Brendon Melone, catSpline is only for categorical data and
uses the softmedian for the harmonised depths to maintain the differential qualities
while aligning with the modal for categorical data.
}
\references{
Malone, B.P., McBratney, A.B., Minasny, B., Laslett, G.M. (2009) Mapping continuous depth functions of soil carbon storage and available water capacity. Geoderma, 154(1-2): 138-152.

Bishop, T.F.A., McBratney, A.B., Laslett, G.M., (1999) Modelling soil attribute depth functions with equal-area quadratic smoothing splines. Geoderma, 91(1-2): 27-45.
}
\author{
Trevan Flynn
}
